Added component for invite card, fixed issue where title wasn't being added to message
