+++ App.js
@@ -16,13 +16,14 @@ export default function App() {
 	<NavigationContainer>
 		<Stack.Navigator initialRoutName="Route" headerMode="false" 
                        screenOptions={{headerShown: false }}>
-			
-			<Stack.Screen name="MainLoginScreen" component={LoginView} />
 			<Stack.Screen
 				name="Dashboard"
 				component={Dashboard}
 				options={{ title: 'Dashboard', headerShown:false }}
 			/>
+
+			<Stack.Screen name="MainLoginScreen" component={LoginView} />
+			
 			<Stack.Screen 
 				name="Messaging"
 				component={Messaging}
